Started demonic gateway, have to fix motion still after clicking the gateway.
Added overried to creature script to allow a script to be assigned to multiple creature ids without adding addtional attributes.

added check to grid to ensure obj is not null, there is some edge case that is adding a null obj here. Added logging to help track it down when it happens.